pyproject.toml
[project]
name = "data_academy_python_day"
version = "0.1.0"
description = "Python day for the data academy"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas==2.3.1",
    "notebook",
    "fastapi[standard]",
    "psycopg2-binary"
]